refactor(bpe): flatten the filename prefixes, group the released-crate BPE under legacy/
