+++ tokenizers/tk-encode/src/models/bpe/mod.rs
@@ -1,15 +1,13 @@
 //! [Byte Pair Encoding](https://www.aclweb.org/anthology/P16-1162/) model.
 use std::{iter, mem};
-mod bpe_build_tables;
-mod bpe_model;
-mod bpe_pretoken_to_rank;
 mod bpe_scratch;
 mod bytelevel_folding;
-mod legacy_model;
-mod legacy_serialization;
-pub mod legacy_word;
+mod convert;
+mod legacy;
 mod merge_hot_cold_queue;
 mod merge_multipass;
+mod model;
+mod tables;
 
 #[cfg(test)]
 mod tests;
@@ -91,7 +89,7 @@ where
 }
 
 // Re-export
-pub use bpe_model::*;
 pub use bpe_scratch::*;
-pub use legacy_model::*;
-pub use legacy_word::*;
+pub use legacy::model::*;
+pub use legacy::word::*;
+pub use model::*;
