+++ tokenizers/tk-encode/src/models/bpe/model.rs
@@ -1,14 +1,14 @@
 //! The pipeline BPE model: its tables, how it is built from a [`BPE`], and how a pretokenized
-//! sequence is turned into tokens. The merge engines themselves live in `bpe_pretoken_to_rank`, `merge_multipass`
-//! and `merge_hot_cold_queue`.
-use crate::models::bpe::bpe_build_tables::BpeTables;
+//! sequence is turned into tokens. Conversion to symbols lives in `convert`; the merge engines
+//! are `merge_multipass` and `merge_hot_cold_queue`.
 use crate::models::bpe::bpe_scratch::BpeScratch;
-use crate::models::bpe::legacy_model::BPE;
+use crate::models::bpe::legacy::model::BPE;
 use crate::models::bpe::merge_hot_cold_queue::{
     MergeScratch, build_byte_to_gate, content_start, two_tier_queue_merge,
 };
 use crate::models::bpe::merge_multipass::merge_multipass;
-use crate::models::bpe::{Error, bpe_build_tables::At};
+use crate::models::bpe::tables::BpeTables;
+use crate::models::bpe::{Error, tables::At};
 use crate::pipeline::{self, PipelineToken};
 use crate::tokenizer::Result;
 use crate::utils::byte_level::{self};
